FlaskBackend/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.1
mysql-connector-python==9.0.0
numpy==2.1.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
